[update] laravel REST API 対応

+++ htdocs/app/Http/Controllers/LearningController.php
@@ -33,30 +33,30 @@ class LearningController extends Controller
    */
   public function learnable($_corpus_id)
   {
-    // try {
-      $corpus = new CorpusManager($_corpus_id);
-      
+    try {
+      $corpus = new CorpusManager($_corpus_id);      
       $formatter = null;
-      if ($corpus->isLearnable()) {
-        $message = 'This corps is learnable status.';
-        $formatter = new ApiResponseFormatter(200, $message, array(
-          'message' => $corpus->getMessage()
-        ));
+      $code = null;
+      $message = null;
+      $data = null;
 
+      if ($corpus->isLearnable()) {
+        $code = 200;
+        $message = 'The specified corpus is learnable.';
+        $data = [array('message' => '学習可能')];
       } else {
-        $message = 'This corpus is not learnale status. -> corpus_id:' . $_corpus_id;
-        $formatter = new ApiResponseFormatter(400, $message, array(
-          'message' => $corpus->getMessage()
-        ));
+        $code = 400;
+        $message = 'Learning can not be performed on the specified corpus.';
+        $data = [array('message' => '学習不可')];
       }
 
-    // } catch (\Exception $e) {
-    //   $message = 'This corpus is not learnale status. ' . $e->getMessage() ;
-    //   $formatter = new ApiResponseFormatter(400, $message, array(
-    //     'message' => 'ご指定のコーパスは学習できません'
-    //   ));
-    // }
+    } catch (\Exception $e) {
+      $code = 400;
+      $message = $e->getMessage();
+      $data = [array('debug' => 'Exception from LearningController@learnable')];
+    }
 
+    $formatter = new ApiResponseFormatter($code, $message, $data);
     return response()->json($formatter->getResponseArray());
   }
 
@@ -67,14 +67,26 @@ class LearningController extends Controller
    */
   public function learn($_corpus_id)
   {
-    $data = new TrainingDataManager($_corpus_id);
-    return $data->deleteAllCsv();
-    // $learnable_response = $this->learnable($_corpus_id);
-    // if ($learnable_response->getData()->code == ApiResponseType::Ok) {
-    //   $nlc = new NaturalLangageClassifierManager($_corpus_id);
-    //   return $nlc;
-    // }
+    // 学習が行える状況かどうかの確認
+    $learnable_response = $this->learnable($_corpus_id);
+    if ($learnable_response->getData()->code == ApiResponseType::Ok) {
+      $corpus = new CorpusManager($_corpus_id);
+      $corpus->execLearn();
 
-    return "pass.";
+      if ($corpus->isErrorExist()) {
+        $formatter = new ApiResponseFormatter(400, $corpus->getMessage(), $corpus->getData());
+      } else {
+        // 正常に学習できた場合は古いCSVファイルを削除して完了コードを返す
+        $data = new TrainingDataManager($_corpus_id);
+        $data->deleteAllCsv();
+        $formatter = new ApiResponseFormatter($data->getCode(), $data->getMessage(), $data->getData());
+      }
+
+      return response()->json($formatter->getResponseArray());
+
+    } else {
+      // 学習出来ない場合はエラーコードをそのまま返す
+      return $learnable_response;
+    }
   }
 }
